+++ src/controller/artifact/model.go
@@ -21,7 +21,6 @@ import (
 	"github.com/goharbor/harbor/src/controller/tag"
 	"github.com/goharbor/harbor/src/lib/encode/repository"
 	"github.com/goharbor/harbor/src/pkg/artifact"
-	"github.com/goharbor/harbor/src/server/v2.0/models"
 )
 
 // Artifact is the overall view of artifact
@@ -50,14 +49,6 @@ func (artifact *Artifact) SetAdditionLink(addition, version string) {
 type AdditionLink struct {
 	HREF     string `json:"href"`
 	Absolute bool   `json:"absolute"` // specify the href is an absolute URL or not
-}
-
-// ToSwagger converts the addition link to the swagger model
-func (a *AdditionLink) ToSwagger() models.AdditionLink {
-	return models.AdditionLink{
-		Absolute: a.Absolute,
-		Href:     a.HREF,
-	}
 }
 
 // Option is used to specify the properties returned when listing/getting artifacts
